+++ src/pep8_naming/helpers.rs
@@ -1,7 +1,8 @@
+use fnv::{FnvHashMap, FnvHashSet};
 use itertools::Itertools;
-use rustpython_ast::{Expr, ExprKind};
+use rustpython_ast::{Expr, ExprKind, Stmt, StmtKind};
 
-use crate::ast::helpers::match_name_or_attr;
+use crate::ast::helpers::{compose_call_path, match_call_path, match_name_or_attr};
 use crate::ast::types::{Scope, ScopeKind};
 use crate::pep8_naming::settings::Settings;
 use crate::python::string::{is_lower, is_upper};
@@ -76,6 +77,19 @@ pub fn is_mixed_case(name: &str) -> bool {
 
 pub fn is_acronym(name: &str, asname: &str) -> bool {
     name.chars().filter(|c| c.is_uppercase()).join("") == asname
+}
+
+pub fn is_namedtuple_assignment(
+    stmt: &Stmt,
+    from_imports: &FnvHashMap<&str, FnvHashSet<&str>>,
+) -> bool {
+    if let StmtKind::Assign { value, .. } = &stmt.node {
+        compose_call_path(value)
+            .map(|call_path| match_call_path(&call_path, "collections.namedtuple", from_imports))
+            .unwrap_or(false)
+    } else {
+        false
+    }
 }
 
 #[cfg(test)]
